==> 03_streets_simplification.ipynb <==
import geopandas as gpd
import neatnet

# load street network and preprocessed buildings of a site
streets = gpd.read_file(r"overture_data\berlin_streets.gpkg")
buildings = gpd.read_file(r"preprocessed_data\berlin_buildings_preprocessed.gpkg")

#streets = gpd.read_file(r"overture_data\hongkong_streets.gpkg")
#buildings = gpd.read_file(r"preprocessed_data\hongkong_buildings_preprocessed.gpkg")

#streets = gpd.read_file(r"overture_data\paris_streets.gpkg")
#buildings = gpd.read_file(r"preprocessed_data\paris_buildings_preprocessed.gpkg")

#streets = gpd.read_file(r"overture_data\rome_streets.gpkg")
#buildings = gpd.read_file(r"preprocessed_data\rome_buildings_preprocessed.gpkg")

#streets = gpd.read_file(r"overture_data\saopaulo_streets.gpkg")
#buildings = gpd.read_file(r"preprocessed_data\saopaulo_buildings_preprocessed.gpkg")

# simplify using neatnet
simplified = neatnet.simplify_network(streets, exclusion_mask=buildings.geometry)

# save
simplified.to_file(r"preprocessed_data\berlin_streets_simplified.gpkg", driver="GPKG")
#simplified.to_file(r"preprocessed_data\hongkong_streets_simplified.gpkg", driver="GPKG")
#simplified.to_file(r"preprocessed_data\paris_streets_simplified.gpkg", driver="GPKG")
#simplified.to_file(r"preprocessed_data\rome_streets_simplified.gpkg", driver="GPKG")
#simplified.to_file(r"preprocessed_data\saopaulo_streets_simplified.gpkg", driver="GPKG")
