==> 04_topology_check.ipynb <==
import geopandas as gpd

# load preprocessed buildings, simplified street network, waterbodies and waterlines of a site
buildings = gpd.read_file(r"preprocessed_data\berlin_buildings_preprocessed.gpkg")
streets = gpd.read_file(r"preprocessed_data\berlin_streets_simplified.gpkg")
waterlines = gpd.read_file(r"overture_data\berlin_waterlines.gpkg")
waterpolygons = gpd.read_file(r"overture_data\berlin_waterbodies.gpkg")

# buildings = gpd.read_file(r"preprocessed_data\hongkong_buildings_preprocessed")
# streets = gpd.read_file(r"preprocessed_data\hongkong_streets_simplified.gpkg")
# waterlines = gpd.read_file(r"overture_data\hongkong_waterlines.gpkg")
# waterpolygons = gpd.read_file(r"overture_data\hongkong_waterbodies.gpkg")

# buildings = gpd.read_file(r"preprocessed_data\paris_buildings_preprocessed")
# streets = gpd.read_file(r"preprocessed_data\paris_streets_simplified.gpkg")
# waterlines = gpd.read_file(r"overture_data\paris_waterlines.gpkg")
# waterpolygons = gpd.read_file(r"overture_data\paris_waterbodies.gpkg")

# buildings = gpd.read_file(r"preprocessed_data\rome_buildings_preprocessed")
# streets = gpd.read_file(r"preprocessed_data\rome_streets_simplified.gpkg")
# waterlines = gpd.read_file(r"overture_data\rome_waterlines.gpkg")
# waterpolygons = gpd.read_file(r"overture_data\rome_waterbodies.gpkg")

# buildings = gpd.read_file(r"preprocessed_data\saopaulo_buildings_preprocessed")
# streets = gpd.read_file(r"preprocessed_data\saopaulo_streets_simplified.gpkg")
# waterlines = gpd.read_file(r"overture_data\saupaulo_waterlines.gpkg")
# waterpolygons = gpd.read_file(r"overture_data\saupaulo_waterbodies.gpkg")

# drop buildings intersecting with streets
intersected_buildings = gpd.sjoin(buildings, streets, how="inner", predicate="intersects")
buildings_checked_streets = buildings.drop(intersected_buildings.index.unique())
buildings_checked_streets = buildings_checked_streets.reset_index(drop=True)

# drop buildings intersecting with waterbodies
intersected_buildings = gpd.sjoin(buildings_checked_streets, waterpolygons, how="inner", predicate="intersects")
buildings_checked = buildings_checked_streets.drop(intersected_buildings.index.unique())
buildings_checked = buildings_checked.reset_index(drop=True)
buildings_checked['building_index'] = buildings_checked.index

# drop waterlines intersecting with buildings
intersected_waterlines = gpd.sjoin(waterlines, buildings_checked_waterbodies, how="inner", predicate="intersects")
waterlines_checked = waterlines.drop(intersected_waterlines.index.unique())
waterlines_checked = waterlines_checked.reset_index(drop=True)

# save
buildings_checked.to_file(r"preprocessed_data\berlin_buildings_checked.gpkg", driver="GPKG")
waterlines_checked.to_file(r"preprocessed_data\berlin_waterlines_checked.gpkg", driver="GPKG")

# buildings_checked.to_file(r"preprocessed_data\hongkong_buildings_checked.gpkg", driver="GPKG")
# waterlines_checked.to_file(r"preprocessed_data\hongkong_waterlines_checked.gpkg", driver="GPKG")

# buildings_checked.to_file(r"preprocessed_data\paris_buildings_checked.gpkg", driver="GPKG")
# waterlines_checked.to_file(r"preprocessed_data\paris_waterlines_checked.gpkg", driver="GPKG")

# buildings_checked.to_file(r"preprocessed_data\rome_buildings_checked.gpkg", driver="GPKG")
# waterlines_checked.to_file(r"preprocessed_data\rome_waterlines_checked.gpkg", driver="GPKG")

# buildings_checked.to_file(r"preprocessed_data\saopaulo_buildings_checked.gpkg", driver="GPKG")
# waterlines_checked.to_file(r"preprocessed_data\saopaulo_waterlines_checked.gpkg", driver="GPKG")
