==> 10_S2_prediction.ipynb <==
import numpy as np
import random
import pandas as pd
import os 
import geopandas as gpd 
from msmla50 import MSMLA50
import gc
import utils
import torch
import cnn_utils
import pickle

torch.manual_seed(0)
np.random.seed(0)
torch.cuda.manual_seed(0)
random.seed(0)
torch.cuda.manual_seed_all(0)
torch.backends.cudnn.benchmark = False
torch.backends.cudnn.deterministic = True
torch.use_deterministic_algorithms(True, warn_only=True)
os.environ["CUBLAS_WORKSPACE_CONFIG"] = ":4096:8"
np.random.default_rng(seed=0)

# settings
patch_size = 32
stride = 10
gt_stride = 32
background_label = 0
batch_size = 128
offset_left = 'best'
offset_top = 'best'

# Berlin

# load splited reference data
splited_ref_data = gpd.read_file(r'ref_data\berlin_ref_splitS2S3S4.gpkg')

# load satellite image (10 m resolution)
image = r'imagery\berlin_20170519.tif'

# load trained S2 CNN models for each fold
cnn_models = [r's2_cnn_models\berlin_S2_fold0_epoch12.pth',
              r's2_cnn_models\berlin_S2_fold1_epoch16.pth',
              r's2_cnn_models\berlin_S2_fold2_epoch22.pth',
              r's2_cnn_models\berlin_S2_fold3_epoch39.pth',
              r's2_cnn_models\berlin_S2_fold4_epoch32.pth']

# define outputs
outputs = [r'outputs\s2\berlin_S2_fold0.tif',
          r'outputs\s2\berlin_S2_fold1.tif',
          r'outputs\s2\berlin_S2_fold2.tif',
          r'outputs\s2\berlin_S2_fold3.tif',
          r'outputs\s2\berlin_S2_fold4.tif']

## Prediction

# recording results
folds = [0, 1, 2, 3, 4]

# results[fold]
results = {
    fold: {} for fold in folds
}

feature_patches = cnn_utils.generate_feature_patches_loader(image_path = image,patch_size = patch_size,stride = stride,batch_size = batch_size,offset_left = offset_left,offset_top = offset_top)

for fold in folds:
    ## prepare training and test polygons
    test_polygons = splited_ref_data[splited_ref_data["fold"] == fold]
    train_polygons = splited_ref_data[splited_ref_data["fold"] != fold]

    train_polygons_raster = fr"berlin_train_f{fold}.tif"
    test_polygons_raster = fr"berlin_test_f{fold}.tif"

    # rasterize
    train_temp = train_polygons_raster.replace(".tif", "_temp.tif")
    test_temp = test_polygons_raster.replace(".tif", "_temp.tif")
    utils.rasterize_reference_polygons(train_polygons, image, train_temp)
    utils.rasterize_reference_polygons(test_polygons, image, test_temp)

    # train and test images matched to 10m image
    train_image_matched = utils.match_rasters(train_temp, image)
    test_image_matched = utils.match_rasters(test_temp, image)

    # save
    train_image_matched.rio.to_raster(train_polygons_raster, driver="GTiff", compress="LZW")
    test_image_matched.rio.to_raster(test_polygons_raster, driver="GTiff", compress="LZW")

    # cleaning
    train_image_matched.close()
    test_image_matched.close()
    train_image_matched = None
    test_image_matched = None
    gc.collect()
    if os.path.exists(train_temp):
        os.remove(train_temp)
    if os.path.exists(test_temp):
        os.remove(test_temp)

    ## load trained CNN model
    cnn_model = MSMLA50(input_channels=10, depth=[16,32,48], num_classes=len(train_polygons["gridcode"].unique()))
    cnn_model = cnn_model.cuda()
    trained_model = torch.load(cnn_models[fold])
    cnn_model.load_state_dict(trained_model['model_state'])

    ## get training patches
    train_patches = cnn_utils.generate_labeled_patches_loader(image_path = image,reference_path = train_polygons_raster,patch_size = patch_size,stride = gt_stride,batch_size = batch_size,offset_left = offset_left,offset_top = offset_top,background_label = background_label)

    ## remapping labels
    label_mapping = cnn_utils.compute_label_mapping(train_patches)
    train_patches = cnn_utils.label_remapping(train_patches, label_mapping)

    ## normalize patches
    mean, std = cnn_utils.get_normalization_parameters(train_patches)
    feature_patches_norm = cnn_utils.normalize_loader(feature_patches, mean, std)

    ## prediction
    print(f"Prediction started for fold {fold}...")
    cnn_model.eval()
    all_preds = list()
    with torch.no_grad():
        for features in feature_patches_norm:
            features = features.cuda()
            
            pred = cnn_model(features)
            pred = pred.cpu()
            
            _, predicted = torch.max(pred, 1)
            all_preds.append(predicted)
    y_pred = torch.cat(all_preds, dim=0)

    ## remapping labels back
    inverse_mapping = {new_label: old_label for old_label, new_label in label_mapping.items()}
    predicted_original_labels = np.array([inverse_mapping[label] for label in y_pred.numpy()])

    ## generate lcz map
    offset_left_calc, offset_top_calc = cnn_utils.calculate_optimal_offsets(image, patch_size, stride)
    utils.lcz_map(offset_left_calc, offset_top_calc, image, predicted_original_labels, outputs[fold])

    print(f"Fold {fold} predicted and saved to {outputs[fold]}.")

## Perpixel validation

# provide test polygons raster path
test_polygons_path = ['berlin_test_f0.tif','berlin_test_f1.tif','berlin_test_f2.tif','berlin_test_f3.tif','berlin_test_f4.tif']

# resample lcz map to 100m
resampled_outputs = []
for f in outputs:
    out = f.replace(".tif", "_100m.tif")
    utils.resample_lcz_map(f, out)
    resampled_outputs.append(out)
    # if os.path.exists(f):
    #     try:
    #         os.remove(f)
    #     except:
    #         pass

metrics, confusion_matrices = utils.perpixel_validation(resampled_outputs, test_polygons_path, splited_ref_data)

df_perpixel = pd.DataFrame(metrics)
df_perpixel = df_perpixel.set_index("Fold")
df_perpixel

df_perpixel_mean = df_perpixel.mean().round(2)
df_perpixel_mean

# export all results to csv
df_perpixel.to_csv(r"results\s2\berlin_S2_results.csv")

# export confusion matrices
with open(r"results\s2\berlin_S2_confusion_matrices.pkl", "wb") as f:
    pickle.dump(confusion_matrices, f)

# Hong Kong

# load splited reference data
splited_ref_data = gpd.read_file(r'ref_data\hongkong_ref_splitS2S3S4.gpkg')

# load satellite image (10 m resolution)
image = r'imagery\hongkong_20180321.tif'

# load trained S2 CNN models for each fold
cnn_models = [r's2_cnn_models\hongkong_S2_fold0_epoch43.pth',
              r's2_cnn_models\hongkong_S2_fold1_epoch38.pth',
              r's2_cnn_models\hongkong_S2_fold2_epoch33.pth',
              r's2_cnn_models\hongkong_S2_fold3_epoch26.pth',
              r's2_cnn_models\hongkong_S2_fold4_epoch29.pth']

# define outputs
outputs = [r'outputs\s2\hongkong_S2_fold0.tif',
          r'outputs\s2\hongkong_S2_fold1.tif',
          r'outputs\s2\hongkong_S2_fold2.tif',
          r'outputs\s2\hongkong_S2_fold3.tif',
          r'outputs\s2\hongkong_S2_fold4.tif']

## Prediction

# recording results
folds = [0, 1, 2, 3, 4]

# results[fold]
results = {
    fold: {} for fold in folds
}

feature_patches = cnn_utils.generate_feature_patches_loader(image_path = image,patch_size = patch_size,stride = stride,batch_size = batch_size,offset_left = offset_left,offset_top = offset_top)

for fold in folds:
    ## prepare training and test polygons
    test_polygons = splited_ref_data[splited_ref_data["fold"] == fold]
    train_polygons = splited_ref_data[splited_ref_data["fold"] != fold]

    train_polygons_raster = fr"hongkong_train_f{fold}.tif"
    test_polygons_raster = fr"hongkong_test_f{fold}.tif"

    # rasterize
    train_temp = train_polygons_raster.replace(".tif", "_temp.tif")
    test_temp = test_polygons_raster.replace(".tif", "_temp.tif")
    utils.rasterize_reference_polygons(train_polygons, image, train_temp)
    utils.rasterize_reference_polygons(test_polygons, image, test_temp)

    # train and test images matched to 10m image
    train_image_matched = utils.match_rasters(train_temp, image)
    test_image_matched = utils.match_rasters(test_temp, image)

    # save
    train_image_matched.rio.to_raster(train_polygons_raster, driver="GTiff", compress="LZW")
    test_image_matched.rio.to_raster(test_polygons_raster, driver="GTiff", compress="LZW")

    # cleaning
    train_image_matched.close()
    test_image_matched.close()
    train_image_matched = None
    test_image_matched = None
    gc.collect()
    if os.path.exists(train_temp):
        os.remove(train_temp)
    if os.path.exists(test_temp):
        os.remove(test_temp)

    ## load trained CNN model
    cnn_model = MSMLA50(input_channels=10, depth=[16,32,48], num_classes=len(train_polygons["gridcode"].unique()))
    cnn_model = cnn_model.cuda()
    trained_model = torch.load(cnn_models[fold])
    cnn_model.load_state_dict(trained_model['model_state'])

    ## get training patches
    train_patches = cnn_utils.generate_labeled_patches_loader(image_path = image,reference_path = train_polygons_raster,patch_size = patch_size,stride = gt_stride,batch_size = batch_size,offset_left = offset_left,offset_top = offset_top,background_label = background_label)

    ## remapping labels
    label_mapping = cnn_utils.compute_label_mapping(train_patches)
    train_patches = cnn_utils.label_remapping(train_patches, label_mapping)

    ## normalize patches
    mean, std = cnn_utils.get_normalization_parameters(train_patches)
    feature_patches_norm = cnn_utils.normalize_loader(feature_patches, mean, std)

    ## prediction
    print(f"Prediction started for fold {fold}...")
    cnn_model.eval()
    all_preds = list()
    with torch.no_grad():
        for features in feature_patches_norm:
            features = features.cuda()
            
            pred = cnn_model(features)
            pred = pred.cpu()
            
            _, predicted = torch.max(pred, 1)
            all_preds.append(predicted)
    y_pred = torch.cat(all_preds, dim=0)

    ## remapping labels back
    inverse_mapping = {new_label: old_label for old_label, new_label in label_mapping.items()}
    predicted_original_labels = np.array([inverse_mapping[label] for label in y_pred.numpy()])

    ## generate lcz map
    offset_left_calc, offset_top_calc = cnn_utils.calculate_optimal_offsets(image, patch_size, stride)
    utils.lcz_map(offset_left_calc, offset_top_calc, image, predicted_original_labels, outputs[fold])

    print(f"Fold {fold} predicted and saved to {outputs[fold]}.")

## Perpixel validation

# provide test polygons raster path
test_polygons_path = ['hongkong_test_f0.tif','hongkong_test_f1.tif','hongkong_test_f2.tif','hongkong_test_f3.tif','hongkong_test_f4.tif']

# resample lcz map to 100m
resampled_outputs = []
for f in outputs:
    out = f.replace(".tif", "_100m.tif")
    utils.resample_lcz_map(f, out)
    resampled_outputs.append(out)
    # if os.path.exists(f):
    #     try:
    #         os.remove(f)
    #     except:
    #         pass

metrics, confusion_matrices = utils.perpixel_validation(resampled_outputs, test_polygons_path, splited_ref_data)

df_perpixel = pd.DataFrame(metrics)
df_perpixel = df_perpixel.set_index("Fold")
df_perpixel

df_perpixel_mean = df_perpixel.mean().round(2)
df_perpixel_mean

# export all results to csv
df_perpixel.to_csv(r"results\s2\hongkong_S2_results.csv")

# export confusion matrices
with open(r"results\s2\hongkong_S2_confusion_matrices.pkl", "wb") as f:
    pickle.dump(confusion_matrices, f)

# Paris

# load splited reference data
splited_ref_data = gpd.read_file(r'ref_data\paris_ref_splitS2S3S4.gpkg')

# load satellite image (10 m resolution)
image = r'imagery\paris_20170526.tif'

# load trained S2 CNN models for each fold
cnn_models = [r's2_cnn_models\paris_S2_fold0_epoch64.pth',
              r's2_cnn_models\paris_S2_fold1_epoch88.pth',
              r's2_cnn_models\paris_S2_fold2_epoch75.pth',
              r's2_cnn_models\paris_S2_fold3_epoch76.pth',
              r's2_cnn_models\paris_S2_fold4_epoch88.pth']

# define outputs
outputs = [r'outputs\s2\paris_S2_fold0.tif',
          r'outputs\s2\paris_S2_fold1.tif',
          r'outputs\s2\paris_S2_fold2.tif',
          r'outputs\s2\paris_S2_fold3.tif',
          r'outputs\s2\paris_S2_fold4.tif']

## Prediction

# recording results
folds = [0, 1, 2, 3, 4]

# results[fold]
results = {
    fold: {} for fold in folds
}

feature_patches = cnn_utils.generate_feature_patches_loader(image_path = image,patch_size = patch_size,stride = stride,batch_size = batch_size,offset_left = offset_left,offset_top = offset_top)

for fold in folds:
    ## prepare training and test polygons
    test_polygons = splited_ref_data[splited_ref_data["fold"] == fold]
    train_polygons = splited_ref_data[splited_ref_data["fold"] != fold]

    train_polygons_raster = fr"paris_train_f{fold}.tif"
    test_polygons_raster = fr"paris_test_f{fold}.tif"

    # rasterize
    train_temp = train_polygons_raster.replace(".tif", "_temp.tif")
    test_temp = test_polygons_raster.replace(".tif", "_temp.tif")
    utils.rasterize_reference_polygons(train_polygons, image, train_temp)
    utils.rasterize_reference_polygons(test_polygons, image, test_temp)

    # train and test images matched to 10m image
    train_image_matched = utils.match_rasters(train_temp, image)
    test_image_matched = utils.match_rasters(test_temp, image)

    # save
    train_image_matched.rio.to_raster(train_polygons_raster, driver="GTiff", compress="LZW")
    test_image_matched.rio.to_raster(test_polygons_raster, driver="GTiff", compress="LZW")

    # cleaning
    train_image_matched.close()
    test_image_matched.close()
    train_image_matched = None
    test_image_matched = None
    gc.collect()
    if os.path.exists(train_temp):
        os.remove(train_temp)
    if os.path.exists(test_temp):
        os.remove(test_temp)

    ## load trained CNN model
    cnn_model = MSMLA50(input_channels=10, depth=[16,32,48], num_classes=len(train_polygons["gridcode"].unique()))
    cnn_model = cnn_model.cuda()
    trained_model = torch.load(cnn_models[fold])
    cnn_model.load_state_dict(trained_model['model_state'])

    ## get training patches
    train_patches = cnn_utils.generate_labeled_patches_loader(image_path = image,reference_path = train_polygons_raster,patch_size = patch_size,stride = gt_stride,batch_size = batch_size,offset_left = offset_left,offset_top = offset_top,background_label = background_label)

    ## remapping labels
    label_mapping = cnn_utils.compute_label_mapping(train_patches)
    train_patches = cnn_utils.label_remapping(train_patches, label_mapping)

    ## normalize patches
    mean, std = cnn_utils.get_normalization_parameters(train_patches)
    feature_patches_norm = cnn_utils.normalize_loader(feature_patches, mean, std)

    ## prediction
    print(f"Prediction started for fold {fold}...")
    cnn_model.eval()
    all_preds = list()
    with torch.no_grad():
        for features in feature_patches_norm:
            features = features.cuda()
            
            pred = cnn_model(features)
            pred = pred.cpu()
            
            _, predicted = torch.max(pred, 1)
            all_preds.append(predicted)
    y_pred = torch.cat(all_preds, dim=0)

    ## remapping labels back
    inverse_mapping = {new_label: old_label for old_label, new_label in label_mapping.items()}
    predicted_original_labels = np.array([inverse_mapping[label] for label in y_pred.numpy()])

    ## generate lcz map
    offset_left_calc, offset_top_calc = cnn_utils.calculate_optimal_offsets(image, patch_size, stride)
    utils.lcz_map(offset_left_calc, offset_top_calc, image, predicted_original_labels, outputs[fold])

    print(f"Fold {fold} predicted and saved to {outputs[fold]}.")

## Perpixel validation

# provide test polygons raster path
test_polygons_path = ['paris_test_f0.tif','paris_test_f1.tif','paris_test_f2.tif','paris_test_f3.tif','paris_test_f4.tif']

# resample lcz map to 100m
resampled_outputs = []
for f in outputs:
    out = f.replace(".tif", "_100m.tif")
    utils.resample_lcz_map(f, out)
    resampled_outputs.append(out)
    # if os.path.exists(f):
    #     try:
    #         os.remove(f)
    #     except:
    #         pass

metrics, confusion_matrices = utils.perpixel_validation(resampled_outputs, test_polygons_path, splited_ref_data)

df_perpixel = pd.DataFrame(metrics)
df_perpixel = df_perpixel.set_index("Fold")
df_perpixel

df_perpixel_mean = df_perpixel.mean().round(2)
df_perpixel_mean

# export all results to csv
df_perpixel.to_csv(r"results\s2\paris_S2_results.csv")

# export confusion matrices
with open(r"results\s2\paris_S2_confusion_matrices.pkl", "wb") as f:
    pickle.dump(confusion_matrices, f)

# Rome

# load splited reference data
splited_ref_data = gpd.read_file(r'ref_data\rome_ref_splitS2S3S4.gpkg')

# load satellite image (10 m resolution)
image = r'imagery\rome_20170620.tif'

# load trained S2 CNN models for each fold
cnn_models = [r's2_cnn_models\rome_S2_fold0_epoch17.pth',
              r's2_cnn_models\rome_S2_fold1_epoch10.pth',
              r's2_cnn_models\rome_S2_fold2_epoch90.pth',
              r's2_cnn_models\rome_S2_fold3_epoch76.pth',
              r's2_cnn_models\rome_S2_fold4_epoch20.pth']

# define outputs
outputs = [r'outputs\s2\rome_S2_fold0.tif',
          r'outputs\s2\rome_S2_fold1.tif',
          r'outputs\s2\rome_S2_fold2.tif',
          r'outputs\s2\rome_S2_fold3.tif',
          r'outputs\s2\rome_S2_fold4.tif']

## Prediction

# recording results
folds = [0, 1, 2, 3, 4]

# results[fold]
results = {
    fold: {} for fold in folds
}

feature_patches = cnn_utils.generate_feature_patches_loader(image_path = image,patch_size = patch_size,stride = stride,batch_size = batch_size,offset_left = offset_left,offset_top = offset_top)

for fold in folds:
    ## prepare training and test polygons
    test_polygons = splited_ref_data[splited_ref_data["fold"] == fold]
    train_polygons = splited_ref_data[splited_ref_data["fold"] != fold]

    train_polygons_raster = fr"rome_train_f{fold}.tif"
    test_polygons_raster = fr"rome_test_f{fold}.tif"

    # rasterize
    train_temp = train_polygons_raster.replace(".tif", "_temp.tif")
    test_temp = test_polygons_raster.replace(".tif", "_temp.tif")
    utils.rasterize_reference_polygons(train_polygons, image, train_temp)
    utils.rasterize_reference_polygons(test_polygons, image, test_temp)

    # train and test images matched to 10m image
    train_image_matched = utils.match_rasters(train_temp, image)
    test_image_matched = utils.match_rasters(test_temp, image)

    # save
    train_image_matched.rio.to_raster(train_polygons_raster, driver="GTiff", compress="LZW")
    test_image_matched.rio.to_raster(test_polygons_raster, driver="GTiff", compress="LZW")

    # cleaning
    train_image_matched.close()
    test_image_matched.close()
    train_image_matched = None
    test_image_matched = None
    gc.collect()
    if os.path.exists(train_temp):
        os.remove(train_temp)
    if os.path.exists(test_temp):
        os.remove(test_temp)

    ## load trained CNN model
    cnn_model = MSMLA50(input_channels=10, depth=[16,32,48], num_classes=len(train_polygons["gridcode"].unique()))
    cnn_model = cnn_model.cuda()
    trained_model = torch.load(cnn_models[fold])
    cnn_model.load_state_dict(trained_model['model_state'])

    ## get training patches
    train_patches = cnn_utils.generate_labeled_patches_loader(image_path = image,reference_path = train_polygons_raster,patch_size = patch_size,stride = gt_stride,batch_size = batch_size,offset_left = offset_left,offset_top = offset_top,background_label = background_label)

    ## remapping labels
    label_mapping = cnn_utils.compute_label_mapping(train_patches)
    train_patches = cnn_utils.label_remapping(train_patches, label_mapping)

    ## normalize patches
    mean, std = cnn_utils.get_normalization_parameters(train_patches)
    feature_patches_norm = cnn_utils.normalize_loader(feature_patches, mean, std)

    ## prediction
    print(f"Prediction started for fold {fold}...")
    cnn_model.eval()
    all_preds = list()
    with torch.no_grad():
        for features in feature_patches_norm:
            features = features.cuda()
            
            pred = cnn_model(features)
            pred = pred.cpu()
            
            _, predicted = torch.max(pred, 1)
            all_preds.append(predicted)
    y_pred = torch.cat(all_preds, dim=0)

    ## remapping labels back
    inverse_mapping = {new_label: old_label for old_label, new_label in label_mapping.items()}
    predicted_original_labels = np.array([inverse_mapping[label] for label in y_pred.numpy()])

    ## generate lcz map
    offset_left_calc, offset_top_calc = cnn_utils.calculate_optimal_offsets(image, patch_size, stride)
    utils.lcz_map(offset_left_calc, offset_top_calc, image, predicted_original_labels, outputs[fold])

    print(f"Fold {fold} predicted and saved to {outputs[fold]}.")

## Perpixel validation

# provide test polygons raster path
test_polygons_path = ['rome_test_f0.tif','rome_test_f1.tif','rome_test_f2.tif','rome_test_f3.tif','rome_test_f4.tif']

# resample lcz map to 100m
resampled_outputs = []
for f in outputs:
    out = f.replace(".tif", "_100m.tif")
    utils.resample_lcz_map(f, out)
    resampled_outputs.append(out)
    # if os.path.exists(f):
    #     try:
    #         os.remove(f)
    #     except:
    #         pass

metrics, confusion_matrices = utils.perpixel_validation(resampled_outputs, test_polygons_path, splited_ref_data)

df_perpixel = pd.DataFrame(metrics)
df_perpixel = df_perpixel.set_index("Fold")
df_perpixel

df_perpixel_mean = df_perpixel.mean().round(2)
df_perpixel_mean

# export all results to csv
df_perpixel.to_csv(r"results\s2\rome_S2_results.csv")

# export confusion matrices
with open(r"results\s2\rome_S2_confusion_matrices.pkl", "wb") as f:
    pickle.dump(confusion_matrices, f)

# Sao Paulo

# load splited reference data
splited_ref_data = gpd.read_file(r'ref_data\saopaulo_ref_splitS2S3S4.gpkg')

# load satellite image (10 m resolution)
image = r'imagery\sao_paulo_20170726.tif'

cnn_models = [r's2_cnn_models\saopaulo_S2_fold0_epoch67.pth',
              r's2_cnn_models\saopaulo_S2_fold1_epoch84.pth',
              r's2_cnn_models\saopaulo_S2_fold2_epoch67.pth',
              r's2_cnn_models\saopaulo_S2_fold3_epoch78.pth',
              r's2_cnn_models\saopaulo_S2_fold4_epoch37.pth']

# define outputs
outputs = [r'outputs\s2\saopaulo_S2_fold0.tif',
          r'outputs\s2\saopaulo_S2_fold1.tif',
          r'outputs\s2\saopaulo_S2_fold2.tif',
          r'outputs\s2\saopaulo_S2_fold3.tif',
          r'outputs\s2\saopaulo_S2_fold4.tif']

## Prediction

# recording results
folds = [0, 1, 2, 3, 4]

# results[fold]
results = {
    fold: {} for fold in folds
}

feature_patches = cnn_utils.generate_feature_patches_loader(image_path = image,patch_size = patch_size,stride = stride,batch_size = batch_size,offset_left = offset_left,offset_top = offset_top)

for fold in folds:
    ## prepare training and test polygons
    test_polygons = splited_ref_data[splited_ref_data["fold"] == fold]
    train_polygons = splited_ref_data[splited_ref_data["fold"] != fold]

    train_polygons_raster = fr"saopaulo_train_f{fold}.tif"
    test_polygons_raster = fr"saopaulo_test_f{fold}.tif"

    # rasterize
    train_temp = train_polygons_raster.replace(".tif", "_temp.tif")
    test_temp = test_polygons_raster.replace(".tif", "_temp.tif")
    utils.rasterize_reference_polygons(train_polygons, image, train_temp)
    utils.rasterize_reference_polygons(test_polygons, image, test_temp)

    # train and test images matched to 10m image
    train_image_matched = utils.match_rasters(train_temp, image)
    test_image_matched = utils.match_rasters(test_temp, image)

    # save
    train_image_matched.rio.to_raster(train_polygons_raster, driver="GTiff", compress="LZW")
    test_image_matched.rio.to_raster(test_polygons_raster, driver="GTiff", compress="LZW")

    # cleaning
    train_image_matched.close()
    test_image_matched.close()
    train_image_matched = None
    test_image_matched = None
    gc.collect()
    if os.path.exists(train_temp):
        os.remove(train_temp)
    if os.path.exists(test_temp):
        os.remove(test_temp)

    ## load trained CNN model
    cnn_model = MSMLA50(input_channels=10, depth=[16,32,48], num_classes=len(train_polygons["gridcode"].unique()))
    cnn_model = cnn_model.cuda()
    trained_model = torch.load(cnn_models[fold])
    cnn_model.load_state_dict(trained_model['model_state'])

    ## get training patches
    train_patches = cnn_utils.generate_labeled_patches_loader(image_path = image,reference_path = train_polygons_raster,patch_size = patch_size,stride = gt_stride,batch_size = batch_size,offset_left = offset_left,offset_top = offset_top,background_label = background_label)

    ## remapping labels
    label_mapping = cnn_utils.compute_label_mapping(train_patches)
    train_patches = cnn_utils.label_remapping(train_patches, label_mapping)

    ## normalize patches
    mean, std = cnn_utils.get_normalization_parameters(train_patches)
    feature_patches_norm = cnn_utils.normalize_loader(feature_patches, mean, std)

    ## prediction
    print(f"Prediction started for fold {fold}...")
    cnn_model.eval()
    all_preds = list()
    with torch.no_grad():
        for features in feature_patches_norm:
            features = features.cuda()
            
            pred = cnn_model(features)
            pred = pred.cpu()
            
            _, predicted = torch.max(pred, 1)
            all_preds.append(predicted)
    y_pred = torch.cat(all_preds, dim=0)

    ## remapping labels back
    inverse_mapping = {new_label: old_label for old_label, new_label in label_mapping.items()}
    predicted_original_labels = np.array([inverse_mapping[label] for label in y_pred.numpy()])

    ## generate lcz map
    offset_left_calc, offset_top_calc = cnn_utils.calculate_optimal_offsets(image, patch_size, stride)
    utils.lcz_map(offset_left_calc, offset_top_calc, image, predicted_original_labels, outputs[fold])

    print(f"Fold {fold} predicted and saved to {outputs[fold]}.")

## Perpixel validation

# provide test polygons raster path
test_polygons_path = ['saopaulo_test_f0.tif','saopaulo_test_f1.tif','saopaulo_test_f2.tif','saopaulo_test_f3.tif','saopaulo_test_f4.tif']

# resample lcz map to 100m
resampled_outputs = []
for f in outputs:
    out = f.replace(".tif", "_100m.tif")
    utils.resample_lcz_map(f, out)
    resampled_outputs.append(out)
    # if os.path.exists(f):
    #     try:
    #         os.remove(f)
    #     except:
    #         pass

metrics, confusion_matrices = utils.perpixel_validation(resampled_outputs, test_polygons_path, splited_ref_data)

df_perpixel = pd.DataFrame(metrics)
df_perpixel = df_perpixel.set_index("Fold")
df_perpixel

df_perpixel_mean = df_perpixel.mean().round(2)
df_perpixel_mean

# export all results to csv
df_perpixel.to_csv(r"results\s2\saopaulo_S2_results.csv")

# export confusion matrices
with open(r"results\s2\saopaulo_S2_confusion_matrices.pkl", "wb") as f:
    pickle.dump(confusion_matrices, f)
